Remove Nick's changes specific to `memory_policy` branch
